Spawn board peaces game objects
There was no board representation yet. Now peaces are generated
in runtime, so the size of the board can be changed in the future.
Each Peace object has its own PeaceComponent (the same happens with
BoardComponent), which is a
Monobehavior object, in order to control Unity routines.

ObjectPoolController is a singleton in order to be accessed from
any script, because it will controll the game.

See: #6